Extract visitor ID management into visitor.ts
